LayoutTests/imported/w3c/web-platform-tests/intersection-observer/v2/box-reflect-expected.txt

Hello, world!

PASS IntersectionObserverV2 detects occlusion from -webkit-box-reflect (if supported).
FAIL First rAF. assert_equals: Occluded if -webkit-box-reflect is supported. expected (boolean) false but got (undefined) undefined
